chat

+++ src/main/java/kr/co/anabada/chat/entity/Chat_Room.java
@@ -17,7 +17,7 @@ public class Chat_Room {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id; // 채팅방 ID
+    private Integer id; // 채팅방 ID (Integer로 수정)
 
     @ManyToOne
     @JoinColumn(name = "seller_id", nullable = false)
@@ -31,7 +31,7 @@ public class Chat_Room {
     private String itemTitle; // 거래 상품 제목
 
     @Column(nullable = false)
-    private Integer itemNo; // 거래 상품 번호 (itemNo 추가)
+    private Integer itemNo; // 거래 상품 번호 (Integer로 수정)
 
     @Builder.Default
     @Column(nullable = false, updatable = false)
